refactor : move InfoPopover.ts to inside InfoPopover/
