+++ Electron/src/componentes/InfoPopover/types/InfoPopover.ts
